softWare2/requirements.txt
azure-functions
azure-storage-blob
wheel
openpyxl
pyecharts==0.1.9.4
xlrd
matplotlib
azure-cognitiveservices-language-textanalytics